requirements.txt
PyQt5==5.15.9
openai>=1.0.0
python-dotenv>=1.0.0